tudo 100%, faltnado apenas o receber doação de recurso

+++ BackEnd/QuataSafeGuardBackEnd/src/main/java/com/quata/quatasafeguardbackend/services/RegistroSaidaItensService.java
@@ -48,10 +48,15 @@ public class RegistroSaidaItensService {
                 .orElseThrow(() -> new NoSuchElementException("Registro de saída não encontrado."));
 
         Produto produto = registro.getProduto();
-        produtoService.adicionarEstoque(produto.getIdProduto(), registro.getQtde());
 
+        int novaQuantidadeEstoque = produto.getQuantidadeEstoque() + registro.getQtde();
+        if (novaQuantidadeEstoque < 0) {
+            throw new IllegalArgumentException("Não é possível restaurar a quantidade de estoque para um valor negativo.");
+        }
+        produtoService.adicionarEstoque(produto.getIdProduto(), registro.getQtde());
         registroSaidaItensRepository.delete(registro);
     }
+
 
     public RegistroSaidaItens atualizarSaida(Long id, Integer novaQuantidade, String novoMotivo) {
         RegistroSaidaItens registro = registroSaidaItensRepository.findById(id)
